requirements.txt
streamlit
torch
transformers
speechbrain
pyannote.audio
ffmpeg-python
librosa
numpy
python-dotenv